Split S3 object handler helpers
